move propagation code to own file `propagation.py`
