+++ openscvx/ptr.py
@@ -7,7 +7,7 @@ import time
 import sys
 from termcolor import colored
 
-from openscvx.discretization import get_discretization_solver, s_to_t, t_to_tau, simulate_nonlinear_time
+from openscvx.propagation import s_to_t, t_to_tau, simulate_nonlinear_time
 from openscvx.config import Config
 from openscvx.ocp import OptimalControlProblem
